scraper/requirements.txt
requests
beautifulsoup4
selenium
chromedriver-autoinstaller
google-generativeai
dotenv